api/requirements.txt
fastapi
uvicorn
firebase-admin
python-dotenv
pydantic
resend
